Move: 도메인에 맞게 클래스 경로 수정
