+++ src/main/java/org/example/sivillage/admin/application/CategoryService.java
@@ -24,20 +24,19 @@ import java.util.stream.Collectors;
 public class CategoryService {
     private final CategoryRepository categoryRepository;
 
+    public void addCategory(AddCategoryRequestDto request) {
+        Category category;
 
-//    public void addCategory(AddCategoryRequestDto request) {
-//        Category category;
-//
-//        if (request.getParentCategoryCode().equals("top")) {
-//            category = Category.createRootCategory(request);
-//        }
-//        else {
-//            Category parentCategory = findCategoryByCategoryCode(request.getParentCategoryCode());
-//            category = Category.createChildCategory(request, parentCategory);
-//        }
-//
-//        categoryRepository.save(category);
-//    }
+        if (request.getParentCategoryCode().equals("top")) {
+            category = Category.createRootCategory(request);
+        }
+        else {
+            Category parentCategory = findCategoryByCategoryCode(request.getParentCategoryCode());
+            category = Category.createChildCategory(request, parentCategory);
+        }
+
+        categoryRepository.save(category);
+    }
 
     public GetSubCategoriesResponseDto getSubCategories(String parentCategoryCode) {
         List<CategoryDto> categories;
@@ -97,7 +96,7 @@ public class CategoryService {
             String categoryName = node.get("categoryName").asText();
             AddCategoryRequestDto requestDto = createRequestDto(categoryName, parentCategory);
 
-            Category category = addCategory(requestDto, parentCategory);
+            Category category = addCategoryWithParentCategory(requestDto, parentCategory);
 
             if (node.has("subCategories")) {
                 for (JsonNode childNode : node.get("subCategories")) {
@@ -107,7 +106,7 @@ public class CategoryService {
         }
     }
 
-    public Category addCategory(AddCategoryRequestDto request, Category parentCategory) {
+    private Category addCategoryWithParentCategory(AddCategoryRequestDto request, Category parentCategory) {
         Category category;
 
         if (parentCategory == null || request.getParentCategoryCode() == null) {
