+++ src/main/java/org/example/sivillage/brand/presentation/BrandController.java
@@ -5,15 +5,16 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.example.sivillage.auth.domain.CustomUserDetails;
 import org.example.sivillage.brand.application.BrandService;
 import org.example.sivillage.brand.dto.in.AddBrandRequestDto;
+import org.example.sivillage.brand.dto.out.GetBrandsListResponseDto;
 import org.example.sivillage.brand.vo.in.AddBrandRequestVo;
+import org.example.sivillage.brand.vo.out.GetBrandsResponseVo;
 import org.example.sivillage.global.common.response.BaseResponse;
 import org.modelmapper.ModelMapper;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.*;
 
 @Slf4j
 @RestController
@@ -31,5 +32,15 @@ public class BrandController {
         brandService.addBrand(request);
 
         return new BaseResponse<>();
+    }
+
+    @Operation(summary = "브랜드 목록 조회")
+    @GetMapping("/")
+    public BaseResponse<GetBrandsResponseVo> getBrands(@AuthenticationPrincipal CustomUserDetails customUserDetails) {
+        GetBrandsListResponseDto getBrandsList = brandService.getBrands(customUserDetails.getMemberUuid());
+        // Mapper를 사용하여 GetBrandslistResponseDto를 GetBrandsResponseVo로 매핑
+        GetBrandsResponseVo response = mapper.map(getBrandsList, GetBrandsResponseVo.class);
+
+        return new BaseResponse<>(response);
     }
 }
